+++ src/main/java/org/datanucleus/management/ManagerStatisticsMBean.java
@@ -13,13 +13,13 @@ public interface ManagerStatisticsMBean
 
     int getQueryExecutionTotalCount();
 
-    int getQueryExecutionTimeLow();
+    long getQueryExecutionTimeLow();
 
-    int getQueryExecutionTimeHigh();
+    long getQueryExecutionTimeHigh();
 
-    int getQueryExecutionTotalTime();
+    long getQueryExecutionTotalTime();
 
-    int getQueryExecutionTimeAverage();
+    long getQueryExecutionTimeAverage();
 
     int getNumberOfDatastoreWrites();
 
@@ -37,13 +37,13 @@ public interface ManagerStatisticsMBean
 
     int getNumberOfObjectDeletes();
 
-    int getTransactionExecutionTimeAverage();
+    long getTransactionExecutionTimeAverage();
 
-    int getTransactionExecutionTimeLow();
+    long getTransactionExecutionTimeLow();
 
-    int getTransactionExecutionTimeHigh();
+    long getTransactionExecutionTimeHigh();
 
-    int getTransactionExecutionTotalTime();
+    long getTransactionExecutionTotalTime();
 
     int getTransactionTotalCount();
